[feat] Rearrange mypage and main page
