+++ src/pages/Main/drawer/index.tsx
@@ -1,0 +1,3 @@
+export const Drawer = () => {
+  return <div>서랍입니다.</div>;
+};
