+++ src/pages/Main/timeTable/index.tsx
@@ -1,0 +1,3 @@
+export const TimeTable = () => {
+  return <div>시간표입니다.</div>;
+};
